requirements.txt
django
django-ckeditor
django-autoslug